refactor(shared): rename asset_object.py to asset_evaluation.py
ユビキタス言語統一の一環。値オブジェクトのファイル名を
概念名そのまま（asset_evaluation.py）に揃える。
クラス名 AssetEvaluation は変更なし。
