+++ lambda/shared/src/shared/domain/asset_evaluation.py
