+++ lambda/summary-notification/src/domain/asset_interface.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from datetime import date
 
-from shared.domain.asset_object import AssetEvaluation
+from shared.domain.asset_evaluation import AssetEvaluation
 
 
 class IAssetRepository(ABC):
