+++ lambda/web-scraping/src/domain/__init__.py
@@ -1,6 +1,6 @@
 """Domain レイヤー: モデル、インターフェース、例外"""
 
-from shared.domain.asset_object import AssetEvaluation
+from shared.domain.asset_evaluation import AssetEvaluation
 from shared.domain.asset_record_interface import IAssetRecordRepository
 from shared.domain.asset_record_object import AssetRecord
 from shared.domain.exceptions import AssetRecordError
